+++ src/Contracts/SocialStrategy.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Webteractive\Passwordless\Contracts;
+
+use Illuminate\Http\Request;
+use Symfony\Component\HttpFoundation\RedirectResponse;
+
+interface SocialStrategy
+{
+    public function redirect(string $provider): RedirectResponse;
+
+    public function callback(string $provider, Request $request): mixed;
+}